requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2024.2.2
click==8.1.7
coverage==7.4.1
Django==3.0
entrypoints==0.3
flake8==3.7.0
gunicorn==21.2.0
h11==0.14.0
iniconfig==2.0.0
mccabe==0.6.1
netcat==0.0.1a0
packaging==23.2
pluggy==1.4.0
psycopg==3.1.18
psycopg2-binary==2.9.9
pycodestyle==2.5.0
pyflakes==2.1.1
pytest==8.0.0
pytest-cov==4.1.0
pytest-django==3.9.0
python-dotenv==1.0.1
pytz==2024.1
sentry-sdk==1.40.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.1
whitenoise==6.6.0
